% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Given a flattened json file, produce R friendly column names.}
\usage{
clean_column_names(json)
}
\arguments{
\item{json}{- JSON file}
}
\value{
data frame
}
\description{
Given a flattened json file, produce R friendly column names.
}
