% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_jurisdiction}
\alias{find_jurisdiction}
\title{Return list of jurisdiction IDs based on the names}
\usage{
find_jurisdiction(name = NA)
}
\arguments{
\item{name}{List of names of the jurisdiction}
}
\value{
List with the jurisdiction IDs of the names specified
}
\description{
Return list of jurisdiction IDs based on the names
}
