% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_agencies}
\alias{get_agencies}
\title{Return a dataframe of all agencies in a jurisdiction}
\usage{
get_agencies(jurisdictionIDs = NULL)
}
\arguments{
\item{jurisdictionIDs}{}
}
\value{
dataframe
}
\description{
Return a dataframe of all agencies in a jurisdiction
}
\examples{
get_agencies(38)
}
