% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_document_types}
\alias{get_document_types}
\title{Title Return the list of document types for a jurisdiction/or all jurisdictions.}
\usage{
get_document_types(jurisdiction = NA)
}
\arguments{
\item{jurisdiction}{Integer - ID for the jurisdiction of interest. If NA, return all jurisdictions}
}
\value{
data frame of document types
}
\description{
Title Return the list of document types for a jurisdiction/or all jurisdictions.
}
\examples{
get_document_types(38)
get_document_types()
}
