% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_documents}
\alias{get_documents}
\title{Title Return the list of documents for a series-jurisdiction combination}
\usage{
get_documents(jurisdiction, document_type = 3)
}
\arguments{
\item{jurisdiction}{Integer - ID of jurisdiction of interest}

\item{document_type}{Integer - ID of document type of interest. Obtain from get_document_types().}
}
\value{
data frame
}
\description{
Title Return the list of documents for a series-jurisdiction combination
}
\examples{
get_documents(55,3)
}
