% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_industries}
\alias{get_industries}
\title{Return a list of industries with data in RegCensus}
\usage{
get_industries(jurisdiction = NA)
}
\arguments{
\item{jurisdiction}{An integer, the jurisdiction ID}
}
\value{
data frame with industries
}
\description{
Return a list of industries with data in RegCensus
}
\examples{
get_industries(38)
}
