% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_industry_values}
\alias{get_industry_values}
\title{Return series values for a set of industry codes (using NAICS)}
\usage{
get_industry_values(
  jurisdiction = c(38),
  industry_type = c("2-Digit", "3-Digit", "4-Digit", "5-Digit", "6-Digit"),
  series = c(1, 9),
  time = c("2015", "2019"),
  agency = c(0),
  date_is_range = TRUE,
  filtered_only = TRUE,
  summary = TRUE,
  document_type = 3
)
}
\arguments{
\item{jurisdiction}{An Integer - jurisdiction(s) of interest. Obtain list of jurisdictions from list_jurisdictions()}

\item{industry_type}{String - List of industry code types desired. Valid values are "all", "2-Digit","3-Digit","4-Digit","5-Digit","6-Digit"}

\item{series}{Integer -  (List of ) Series of interest. Obtain valid list from list_series(id, by)}

\item{time}{Date (string) - String format of dates. For summary data, just the year is enough. For daily data, use full date format such as
'2018-10-12'}

\item{agency}{Integer - List of agencies. Obtain from get_agencies()}

\item{date_is_range}{Boolean - Date parameter is range}

\item{filtered_only}{Boolean - For industry values, include only good-performing industry classifications.}

\item{summary}{Boolean - Summary data}

\item{document_type}{Integer - The type of document. Obtain from list_document_types()}
}
\value{
Data frame
}
\description{
Return series values for a set of industry codes (using NAICS)
}
