% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_jurisdictions}
\alias{get_jurisdictions}
\title{Generate a list of jurisdictions in RegData}
\usage{
get_jurisdictions(jurisdiction = NA)
}
\arguments{
\item{jurisdiction}{Integer - the ID of the jurisdiction of interest}
}
\value{
List of jurisdictions
}
\description{
Generate a list of jurisdictions in RegData
}
\examples{
get_jurisdictions()
get_jurisdictions(38)
}
