% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_series}
\alias{get_series}
\title{Return a dataframe with the series for the id specified}
\usage{
get_series(id = NA)
}
\arguments{
\item{id}{Integer - ID for the by parameter}
}
\value{
dataframe
}
\description{
Return a dataframe with the series for the id specified
}
\examples{
get_series(bidy=1)
}
