% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_series_period}
\alias{get_series_period}
\title{Return dataframe with all jurisdictions-series-years of data available}
\usage{
get_series_period(jurisdiction = NA)
}
\arguments{
\item{jurisdiction}{Integer - the ID of the jurisdiction of interest}
}
\value{
dataframe
}
\description{
Return dataframe with all jurisdictions-series-years of data available
}
\examples{
get_periods()
get_periods(38)
}
