% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_seriesyear}
\alias{get_seriesyear}
\title{Return dataframe with all jurisdictions-series-years of data available}
\usage{
get_seriesyear(jurisdictionID = NA)
}
\arguments{
\item{jurisdictionID}{}
}
\value{
dataframe
}
\description{
Return dataframe with all jurisdictions-series-years of data available
}
\examples{
get_seriesyear()
get_seriesyear(38)
}
