% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R, R/list.R
\name{list_document_types}
\alias{list_document_types}
\title{Title Return the list of document types for a jurisdiction/or all jurisdictions.}
\usage{
list_document_types()

list_document_types()
}
\arguments{
\item{jurisdiction}{Integer - ID for the jurisdiction of interest. If NA, return all jurisdictions}
}
\value{
data frame of document types

Data frame with the list of document types available
}
\description{
Title Return the list of document types for a jurisdiction/or all jurisdictions.

List the document types available for a jurisdiction or all jurisdiction
}
\examples{
list_document_types(38)
list_document_types()
list_document_types()
}
