% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_jurisdictions}
\alias{list_jurisdictions}
\title{Generate a list of Jurisdictions in RegData. This function is helpful for identifying the jurisdictions covered and their identifiers used in the RegData system. Note that the function accepts one jurisdictionID, distinct from FIPs or geoCodes.}
\usage{
list_jurisdictions(jurisdictionID = NA)
}
\arguments{
\item{jurisdictionID}{An integer.}
}
\value{
List of jurisdictions and their corresponding IDs and geoCodes in the console.
}
\description{
Generate a list of Jurisdictions in RegData. This function is helpful for identifying the jurisdictions covered and their identifiers used in the RegData system. Note that the function accepts one jurisdictionID, distinct from FIPs or geoCodes.
}
\examples{
list_jurisdictions()
list_jurisdictions(38)
}
