% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_series}
\alias{list_series}
\title{Generate a list of Series in RegData and their corresponding series IDs and series codes.}
\usage{
list_series(id = NA, by = c("all", "series", "agencies", "industries",
  "jurisdictions", "topics"))
}
\arguments{
\item{id}{An integer or vector of integers. See description above and examples below for when vectors are allowed and when single values are required.}

\item{by}{A string specifying what type of identifier you are trying to pass in the "id" parameter. Default is "series," and accepted values are "seriesCodes", "agencies", "jurisdictions", and "topic".}
}
\value{
List of series and their corresponding IDs and series codes in the console.
}
\description{
You can obtain information on series by:
\itemize{
\item series ID (default). Only one series ID is accepted. Lists by series do not require an id to be defined.
\item series code(s). Multiple series codes can be selected by passing a vector of strings as the "id" parameter.
\item agency or agencies. The function accepts multiple agency IDs passed as a vector.
\item jurisdictions. The function accepts jurisdiction IDs (distinct from geoCodes) as a vector.
\item topic. Only one topic ID is allowed.
}
No "id" is required when selecting series by series. See first example. When selecting series by series codes, agencies, jurisdictions, or topic, the "id" must be specified.
}
\examples{
list_series()
list_series(1)
list_series(1, by="series")
list_series("1", by="series")
list_series(c(91, 92, 100), by="series")
list_series(c(244, 216, 206, 189), by="agencies")
list_series(c(38, 10), by="jurisdictions")
list_series(1, by="topic")
}
