% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_seriesyear}
\alias{list_seriesyear}
\title{Generate a list of Start Years, End Years, and Total Years Available for Series by Jurisdiction in RegData.}
\usage{
list_seriesyear(jurisdiction = NA)
}
\arguments{
\item{jurisdiction}{An integer. Only one jurisdiction value may be passed as an argument.}
}
\value{
List of series with information for each series regarding start year, end years, and total years available, by jurisdiction
}
\description{
Each item in the list returns:
\itemize{
\item Start Year
\item End Year
\item Number of Years Available
\item Series ID
\item Series Code
\item Jurisdiction ID
\item Geo Code
}
}
\examples{
list_seriesyear(10)
}
