% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_topics}
\alias{list_topics}
\title{Generate a list of Topics in RegData}
\usage{
list_topics(topicID = NA)
}
\arguments{
\item{topicID}{An integer.}
}
\value{
List of topics and their corresponding IDs in the console.
}
\description{
Generate a list of Topics in RegData
}
\examples{
list_topics()
list_topics(1)
}
