% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_api_call}
\alias{make_api_call}
\title{Manage API call, including error messages}
\usage{
make_api_call(url, verbose = FALSE)
}
\arguments{
\item{url}{- Text - URL with data.}

\item{verbose}{- Boolean - if true, show URL calls to API}
}
\value{
json with results
}
\description{
Manage API call, including error messages
}
