% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{regdata_json_request}
\alias{regdata_json_request}
\title{Return the basic URL for an API call.}
\usage{
regdata_json_request(controller, id)
}
\arguments{
\item{controller}{The name of the controller}

\item{id}{ID List of IDs to pass to API call}
}
\value{
String
}
\description{
Return the basic URL for an API call.
}
