% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{regdata_json_requestURL}
\alias{regdata_json_requestURL}
\title{Return URL of API Call from list of IDs}
\usage{
regdata_json_requestURL(controller, id)
}
\arguments{
\item{controller}{short name of controller}

\item{id}{List of IDs}
}
\value{
String with URL
}
\description{
Return URL of API Call from list of IDs
}
