% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_country_values}
\alias{get_country_values}
\title{Return values for a list of jurisdictions-series-year combination. It will return data for both
national and sub-national (for example, state/province) within the jurisdiction.}
\usage{
get_country_values(
  jurisdiction = c(38, 75),
  series = c(1, 2),
  date = c(2010, 2011),
  industries = NA,
  date_is_range = TRUE,
  document_type = 3,
  filtered = TRUE
)
}
\arguments{
\item{jurisdiction}{Integer - Jurisdiction IDs for the data}

\item{series}{Integer - Series IDs}

\item{date}{Date string - For period of interest}

\item{industries}{Text - List of industries to pull data for using the jurisdiction specific coding system}

\item{date_is_range}{Boolean - Indicate if the years values is range instead of just single values}

\item{document_type}{Integer - ID of document type}

\item{filtered}{Boolean - Indicate whether to return filtered data, i.e., the industry-relevant data with poorly classified industries excluded}
}
\value{
dataframe
}
\description{
Return values for a list of jurisdictions-series-year combination. It will return data for both
national and sub-national (for example, state/province) within the jurisdiction.
}
