% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_series}
\alias{get_series}
\title{Return a dataframe with the series for the id specified}
\usage{
get_series(id = NA, by = c("all", "series", "agencies", "industries",
  "jurisdictions", "topics"))
}
\arguments{
\item{id}{Integer - ID for the by parameter}

\item{by}{Text - one of all, series, agencies, industries, jurisdictions, topics}
}
\value{
dataframe
}
\description{
Return a dataframe with the series for the id specified
}
\examples{
get_series(by='all')
get_series(id=91, by = 'series')
}
