% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_topics}
\alias{get_topics}
\title{Return list of topics in RegCensus}
\usage{
get_topics(topicID = NA)
}
\arguments{
\item{topicID}{}
}
\value{
dataframe
}
\description{
Return list of topics in RegCensus
}
\examples{
get_topics(1)
get_topics()
}
