% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frames.R
\name{get_values}
\alias{get_values}
\title{Return dataframe with values of series of interest}
\usage{
get_values(
  jurisdiction,
  series,
  time,
  agency = NA,
  industry = NA,
  dateIsRange = TRUE
)
}
\arguments{
\item{dateIsRange}{}
}
\value{
dataframe
}
\description{
Return dataframe with values of series of interest
}
\examples{
get_values(jurisdiction = 38, series = c(92), time = c(1990,2000),
industry = c('111','33'), agency = c(66,111))
}
