% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_document_types}
\alias{list_document_types}
\title{List the document types available for a jurisdiction or all jurisdiction}
\usage{
list_document_types()
}
\value{
Data frame with the list of document types available
}
\description{
List the document types available for a jurisdiction or all jurisdiction
}
\examples{
list_document_types()
}
